--- src/App.scss
@@ -1,17 +1,0 @@
-.cardSection {
-    display: flex;
-    justify-content: center;
-    padding: 24px 0;
-    background: #f7f7f7;
-}
-
-.cardTitle {
-    text-align: center;
-    margin-bottom: 32px;
-}
-
-.cardWrap {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    margin-bottom: 12px;
-}
