+++ src/components/Layout/Layout.module.scss
@@ -1,0 +1,3 @@
+.root {
+
+}
