+++ src/pages/Main/Main.module.scss
@@ -1,0 +1,19 @@
+.root {
+}
+.cardSection {
+    display: flex;
+    justify-content: center;
+    padding: 24px 0;
+    background: #f7f7f7;
+}
+
+.cardTitle {
+    text-align: center;
+    margin-bottom: 32px;
+}
+
+.cardWrap {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    margin-bottom: 12px;
+}
